stack/provider: add generic template tests
