+++ go/src/koding/kites/kloud/stack/provider/testdata/aws-multiple-vms.json.golden
@@ -10,7 +10,7 @@
 		"aws_instance": {
 			"aws-instance": {
 				"ami": "ami-123456",
-				"custom_data": "***",
+				"count": 3,
 				"instance_type": "t2.nano",
 				"key_name": "koding-kp",
 				"security_groups": [
@@ -22,14 +22,16 @@
 						"Name": "${var.koding_user_username}-${var.koding_group_slug}"
 					}
 				],
-				"user_data": "# let's create a file on your root folder:\necho \"hello world!\" \u003e\u003e /helloworld.txt\n# please note: all commands under user_data will be run as root.\n# now add your credentials and save this stack.\n# once vm finishes building, you can see this file by typing\n# ls /\n#\n# for more information please use the search box above"
+				"user_data": "***"
 			}
 		}
 	},
 	"variable": {
 		"kitekeys_aws-instance": {
 			"default": {
-				"0": "***"
+				"0": "***",
+				"1": "***",
+				"2": "***"
 			}
 		}
 	}
