+++ go/src/koding/kites/kloud/stack/provider/testdata/aws-single-vm.json
