+++ go/src/koding/kites/kloud/stack/provider/testdata/azure-basic-stack.json
