+++ go/src/koding/kites/kloud/stack/provider/testdata/azure-custom-endpoint.json
